chore: move EventGenerator to data/event since it's not a use case
